Fixing some things related to Quote and JumpEqualFunction
