+++ engine/src/system/core/model/Remappable.java
@@ -2,7 +2,7 @@ package system.core.model;
 
 import java.util.function.UnaryOperator;
 
-/** Minimal contract so QUOTE can rename vars/labels generically, no big switch. */
+/** Minimal contract so QUOTE can rename vars/labels generically. */
 public interface Remappable {
     /** Return a copy of this instruction after applying varMap and labelMap. */
     Instruction remap(UnaryOperator<Var> varMap,
